# PERF: DataFrame dict constructor with columns