--- a/pandas/tests/frame/test_constructors.py
+++ b/pandas/tests/frame/test_constructors.py
@@ -1401,6 +1401,11 @@ class TestDataFrameConstructors(TestData):
         pytest.raises(ValueError, DataFrame.from_dict,
                       OrderedDict([('b', 8), ('a', 5), ('a', 6)]))
 
+    def test_constructor_column_dict_duplicates(self):
+        result = DataFrame({"A": [1, 2], "B": [3, 4]}, columns=['A', 'B', 'A'])
+        expected = DataFrame({"A": [1, 2], "B": [3, 4]}, columns=['A', 'B'])
+        tm.assert_frame_equal(result, expected)
+
     def test_constructor_empty_with_string_dtype(self):
         # GH 9428
         expected = DataFrame(index=[0, 1], columns=[0, 1], dtype=object)