@pep8speaks 2018-12-21
Hello @TomAugspurger! Thanks for submitting the PR.

 - There are no PEP8 issues in the file [`asv_bench/benchmarks/frame_ctor.py`](https://github.com/TomAugspurger/pandas/blob/a7606e7fd1bc5a54ce2979fb2d5a83b256c9bb0a/asv_bench/benchmarks/frame_ctor.py) !

 - There are no PEP8 issues in the file [`pandas/core/internals/construction.py`](https://github.com/TomAugspurger/pandas/blob/a7606e7fd1bc5a54ce2979fb2d5a83b256c9bb0a/pandas/core/internals/construction.py) !

 - There are no PEP8 issues in the file [`pandas/tests/frame/test_constructors.py`](https://github.com/TomAugspurger/pandas/blob/a7606e7fd1bc5a54ce2979fb2d5a83b256c9bb0a/pandas/tests/frame/test_constructors.py) !