@TomAugspurger 2018-12-21
Note that this passes the tests in `tests/frame/test_constructors.py`, but other tests are failing. I'll be updating to handle additional edge cases that apparently aren't tested in constructors.