@TomAugspurger 2018-12-21
With the latest all of pandas/tests/frame is passing...