@TomAugspurger 2019-03-05
Yep.

On Wed, Feb 27, 2019 at 5:56 PM William Ayd <notifications@github.com>
wrote:

> @TomAugspurger <https://github.com/TomAugspurger> still active?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/pandas-dev/pandas/pull/24387#issuecomment-468079100>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABQHIkbvW5XPdn2pRl6IqcqyHtXdLkS-ks5vRxspgaJpZM4Zezaf>
> .
>