@jreback 2019-03-26
pushing the benchmarks here is probably worth it. other parts would need to be revisited. closing.